Allowed administrators to call javascript report generators to publish master report parts: MasterSummary, FailureSummary, and OnSummary (for local report customization, currently just outputs apis called).
Fixed bug where when deployed on Java 8, would use the default JS of Nashorn, but now we default to Rhino by jar inclusion. This preserves the syntax that Rhino java objects can have their static methods call on their instance from the javascript side.  This syntax is used in the Tools and Kit class and is considered "normal".

Fixed a bug where masterVars were cloned, but not passes to the eval in readHeaders().  Now mastervars will get passed all the way into headers with expressions.
remove me